# Release v1.2.5: Performance optimization (+10-13% throughput)