--- a/oxidize-pdf-core/src/dashboard/pivot_table.rs
+++ b/oxidize-pdf-core/src/dashboard/pivot_table.rs
@@ -81,44 +81,184 @@ impl PivotTable {
 impl DashboardComponent for PivotTable {
     fn render(
         &self,
-        _page: &mut Page,
+        page: &mut Page,
         position: ComponentPosition,
-        _theme: &DashboardTheme,
+        theme: &DashboardTheme,
     ) -> Result<(), PdfError> {
         let mut table = self.clone();
         table.ensure_computed()?;
 
         let computed = table.computed_data.as_ref().unwrap();
 
+        if computed.headers.is_empty() {
+            return Ok(());
+        }
+
+        let title_height = if table.pivot_config.title.is_some() {
+            30.0
+        } else {
+            0.0
+        };
+        let row_height = 22.0;
+        let header_height = 25.0;
+        let padding = 5.0;
+
+        let mut current_y = position.y + position.height - title_height;
+
         // Render title if present
-        if let Some(ref _title) = table.pivot_config.title {
-            // Placeholder: page.add_text replaced
+        if let Some(ref title) = table.pivot_config.title {
+            page.text()
+                .set_font(crate::Font::HelveticaBold, theme.typography.heading_size)
+                .set_fill_color(theme.colors.text_primary)
+                .at(position.x, current_y - 15.0)
+                .write(title)?;
+            current_y -= title_height;
         }
 
-        // Simple table rendering (placeholder)
-        let mut current_y = position.y + position.height - 40.0;
-        let row_height = 20.0;
+        // Calculate column widths
+        let col_width = position.width / computed.headers.len() as f64;
 
-        // Render headers
-        for (i, _header) in computed.headers.iter().enumerate() {
-            let _x = position.x + i as f64 * (position.width / computed.headers.len() as f64);
-            // Placeholder: page.add_text replaced
+        // Render header row with background
+        page.graphics()
+            .set_fill_color(crate::graphics::Color::gray(0.9))
+            .rect(
+                position.x,
+                current_y - header_height,
+                position.width,
+                header_height,
+            )
+            .fill();
+
+        // Render header border
+        page.graphics()
+            .set_stroke_color(crate::graphics::Color::gray(0.6))
+            .set_line_width(1.0)
+            .rect(
+                position.x,
+                current_y - header_height,
+                position.width,
+                header_height,
+            )
+            .stroke();
+
+        // Render header text
+        for (i, header) in computed.headers.iter().enumerate() {
+            let x = position.x + i as f64 * col_width + padding;
+
+            page.text()
+                .set_font(crate::Font::HelveticaBold, 10.0)
+                .set_fill_color(theme.colors.text_primary)
+                .at(x, current_y - header_height + 7.0)
+                .write(header)?;
+
+            // Draw column separator
+            if i < computed.headers.len() - 1 {
+                let sep_x = position.x + (i + 1) as f64 * col_width;
+                page.graphics()
+                    .set_stroke_color(crate::graphics::Color::gray(0.6))
+                    .set_line_width(0.5)
+                    .move_to(sep_x, current_y - header_height)
+                    .line_to(sep_x, current_y)
+                    .stroke();
+            }
         }
 
-        current_y -= row_height;
+        current_y -= header_height;
 
         // Render data rows
         for (row_idx, row) in computed.rows.iter().enumerate() {
             let is_totals = computed.totals_row == Some(row_idx);
 
-            for (col_idx, _cell) in row.iter().enumerate() {
-                let _x =
-                    position.x + col_idx as f64 * (position.width / computed.headers.len() as f64);
-                let _is_totals = is_totals; // Suppress warning
-                                            // Placeholder: page.add_text replaced
+            // Alternate row background
+            if !is_totals && row_idx % 2 == 1 {
+                page.graphics()
+                    .set_fill_color(crate::graphics::Color::gray(0.97))
+                    .rect(
+                        position.x,
+                        current_y - row_height,
+                        position.width,
+                        row_height,
+                    )
+                    .fill();
             }
+
+            // Totals row background
+            if is_totals {
+                page.graphics()
+                    .set_fill_color(crate::graphics::Color::gray(0.85))
+                    .rect(
+                        position.x,
+                        current_y - row_height,
+                        position.width,
+                        row_height,
+                    )
+                    .fill();
+            }
+
+            // Draw row border
+            page.graphics()
+                .set_stroke_color(crate::graphics::Color::gray(0.8))
+                .set_line_width(0.5)
+                .move_to(position.x, current_y - row_height)
+                .line_to(position.x + position.width, current_y - row_height)
+                .stroke();
+
+            // Render cells
+            for (col_idx, cell) in row.iter().enumerate() {
+                let x = position.x + col_idx as f64 * col_width + padding;
+
+                let font = if is_totals {
+                    crate::Font::HelveticaBold
+                } else {
+                    crate::Font::Helvetica
+                };
+
+                page.text()
+                    .set_font(font, 9.0)
+                    .set_fill_color(theme.colors.text_primary)
+                    .at(x, current_y - row_height + 6.0)
+                    .write(cell)?;
+
+                // Draw column separator
+                if col_idx < row.len() - 1 {
+                    let sep_x = position.x + (col_idx + 1) as f64 * col_width;
+                    page.graphics()
+                        .set_stroke_color(crate::graphics::Color::gray(0.8))
+                        .set_line_width(0.5)
+                        .move_to(sep_x, current_y - row_height)
+                        .line_to(sep_x, current_y)
+                        .stroke();
+                }
+            }
+
             current_y -= row_height;
         }
+
+        // Draw final bottom border
+        page.graphics()
+            .set_stroke_color(crate::graphics::Color::gray(0.6))
+            .set_line_width(1.0)
+            .move_to(position.x, current_y)
+            .line_to(position.x + position.width, current_y)
+            .stroke();
+
+        // Draw left and right borders
+        page.graphics()
+            .set_stroke_color(crate::graphics::Color::gray(0.6))
+            .set_line_width(1.0)
+            .move_to(position.x, position.y + position.height - title_height)
+            .line_to(position.x, current_y)
+            .stroke();
+
+        page.graphics()
+            .set_stroke_color(crate::graphics::Color::gray(0.6))
+            .set_line_width(1.0)
+            .move_to(
+                position.x + position.width,
+                position.y + position.height - title_height,
+            )
+            .line_to(position.x + position.width, current_y)
+            .stroke();
 
         Ok(())
     }