… PNG_DECODER_ISSUES.md
… PROJECT_PROGRESS.md
… ROADMAP.md
… examples/src/charts_dashboard_integration_test.rs
… examples/src/dashboard_sales_report.rs
… examples/src/dashboard_templates_demo.rs
… examples/src/extract_images_demo.rs
… examples/src/heatmap_simple_test.rs
… examples/src/line_chart_dashboard_test.rs
… examples/src/performance_benchmark_1000.rs
… examples/src/pivot_table_simple_test.rs
… examples/src/scatter_plot_simple_test.rs
… examples/src/treemap_simple_test.rs
… oxidize-pdf-core/Cargo.toml
… oxidize-pdf-core/src/dashboard/heatmap.rs
… oxidize-pdf-core/src/dashboard/scatter_plot.rs
… oxidize-pdf-core/src/dashboard/templates.rs
… oxidize-pdf-core/src/dashboard/treemap.rs
… oxidize-pdf-core/src/document.rs
… oxidize-pdf-core/src/graphics/pdf_image.rs
… oxidize-pdf-core/src/operations/page_analysis_ocr_tests.rs
… oxidize-pdf-core/src/page.rs
… oxidize-pdf-core/src/parser/page_tree.rs
… oxidize-pdf-core/src/parser/reader.rs
… oxidize-pdf-core/src/parser/stack_safe.rs
… oxidize-pdf-core/src/parser/stack_safe_tests.rs
… oxidize-pdf-core/src/parser/xref.rs
… oxidize-pdf-core/src/parser/xref_stream.rs
… oxidize-pdf-core/src/text/ocr.rs
… oxidize-pdf-core/src/text/tesseract_provider.rs